finish add service layer

+++ src/Service/EquipmentService.php
@@ -5,6 +5,8 @@ namespace App\Service;
 use App\Classes\Constants;
 use App\Entity\Equipment;
 use App\Repository\EquipmentRepository;
+use App\Repository\UserRepository;
+use App\Repository\AssignRepository;
 
 use App\Form\EquipmentType;
 use Symfony\Component\HttpFoundation\Request;
@@ -12,16 +14,21 @@ use Symfony\Component\HttpFoundation\Session\Session;
 use Symfony\Component\HttpFoundation\RequestStack;
 use Symfony\Component\Validator\Validator\ValidatorInterface;
 use Symfony\Component\HttpFoundation\Response;
+use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 
 
-class EquipmentService {
-
+class EquipmentService extends AbstractController
+{
+    private $assignRepository;
     private $equipmentRepository;
+    private $userRepository;
     private $session;
     private $validator; 
     
-    public function __construct(EquipmentRepository $equipmentRepository,RequestStack  $session,ValidatorInterface $validator) {
+    public function __construct(UserRepository $userRepository, AssignRepository $assignRepository,EquipmentRepository $equipmentRepository,RequestStack  $session,ValidatorInterface $validator) {
         $this->equipmentRepository = $equipmentRepository;
+        $this->assignRepository = $assignRepository;
+        $this->userRepository = $userRepository;
         $this->session = $session->getSession();
         $this->validator = $validator;
     }
@@ -30,10 +37,16 @@ class EquipmentService {
         return $this->equipmentRepository->getAll();
     }
 
+    public function findOne($id) {
+        return $this->equipmentRepository->findOne();
+    }
+
     public function create(Request $request,$form) {
         $form->handleRequest($request);
         if ($form->isSubmitted() && $form->isValid()) {
             $equipment = $form->getData();
+            $this->equipmentRepository->setStatus($equipment,constants::STATUS_AVAILABLE);
+            
             $errors = $this->validator->validate($equipment);
             
             if (count($errors) > 0) {
@@ -50,7 +63,7 @@ class EquipmentService {
     }
 
 
-    public function edit(Request $request,$form) {
+    public function edit(Request $request,$form) { 
         $form->handleRequest($request);
         if ($form->isSubmitted() && $form->isValid()) {
             $equipment = $form->getData();
@@ -70,8 +83,33 @@ class EquipmentService {
         return false;
     }
 
-    public function delete($id) {
-        $this->equipmentRepository->delete($id,true);
-        $this->session->set('success','Equipment deleted');
+    public function delete($id,Request $request) {
+        if ($this->isCsrfTokenValid('delete', $request->request->get('_token'))) {
+            $this->equipmentRepository->delete($id,true);
+            $this->session->set('success','Equipment deleted');
+        }
+        else $this->session->set('failed','Unable to delete');
+    }
+
+    public function assign($id, Request $request) {
+        if ($this->isCsrfTokenValid('assign', $request->request->get('_token'))) {
+            $equipment = $this->equipmentRepository->findOne($id);
+            $user_id = $request->request->get('user_id');
+            $user = $this->userRepository->findOne($user_id);
+
+            $this->assignRepository->store($user,$equipment);
+            $this->equipmentRepository->setStatus($equipment,Constants::STATUS_IN_USE);
+            $this->session->set('success','Equipment assigned');
+        }
+        else $this->session->set('failed','Unable to assign');
+    }
+
+    public function unassign($id, Request $request) {
+        if ($this->isCsrfTokenValid('unassign', $request->request->get('_token'))) {
+            $equipment = $this->equipmentRepository->findOne($id);
+            $this->equipmentRepository->setStatus($equipment,Constants::STATUS_AVAILABLE);
+            $this->session->set('success','Equipment unassigned');
+        }
+        else $this->session->set('failed','Unable to unassign');
     }
 }
